Extract transactions commands and tests to separate files
